Health "probes" working
