+++ 2024/workspaces/apps/proxy/nginx.conf
@@ -9,6 +9,8 @@ http {
     # https://stackoverflow.com/a/58008089
     # https://nickjanetakis.com/blog/using-nginx-regex-capture-groups-to-redirect-url-paths
 
+    # TODO: proxy the REST API under /_api?
+
     # https://stackoverflow.com/a/72569511
     location ~* /workspaces/(.*)/(\d+)/(.*) {#~ ^/workspaces/([\w\-]+)/(\d+)$ {
       # https://serverfault.com/a/1100500
