+++ 2024/workspaces/apps/server/server/routes/well_known.py
@@ -1,0 +1,9 @@
+from fastapi import APIRouter
+
+
+router = APIRouter()
+
+
+@router.get("/healthz")
+def healthz():
+    return "ok"
