+++ 2024/workspaces/tests/test_workspace_api.py
@@ -15,7 +15,7 @@ def test_get_workspace(api, workspace):
     api.get(f"/workspaces/{workspace['id']}").raise_for_status()
 
 
-@pytest.mark.timeout(60)
+@pytest.mark.timeout(30)
 def test_wait_for_workspace_to_be_ready(api, workspace):
     status = ""
 
@@ -30,13 +30,15 @@ def test_wait_for_workspace_to_be_ready(api, workspace):
 
 
 # TODO: move this "into" workspace-specific health probe.
-@pytest.mark.timeout(30)
+@pytest.mark.timeout(10)
 def test_wait_for_workspace_available_via_proxy(proxy, health):
     success = False
     while not success:
         time.sleep(3)
-        success = proxy.get(health).is_success
-        print(f"{success=}")
+        response = proxy.get(health)
+        success = response.is_success
+        print(f"{response.request.url} {response.status_code=} {response.text=}")
+    response.raise_for_status()
 
 
 def test_delete_workspace(api, workspace):
@@ -48,7 +50,7 @@ def test_workspace_no_longer_exists(api, workspace):
 
 
 def test_delete_dangling_workspaces(api):
-    workspaces = api.get("/workspaces").raise_for_status().json()
+    workspaces = api.get("/workspaces/").raise_for_status().json()
     responses = []
     for workspace in workspaces:
         responses.append(api.delete(f"/workspaces/{workspace['id']}"))
